Adapter for RecyclerView was added

+++ app/src/main/java/com/example/todolist/entities/Note.kt
@@ -10,6 +10,6 @@ import org.jetbrains.annotations.NotNull
 data class Note(
     @PrimaryKey val id: Int,
     @NotNull
-    @ColumnInfo val creationTime : Int,
+    @ColumnInfo val creationTime : Long,
     @NotNull
     @ColumnInfo var body : String)
